pyproject.toml
[project]
name = "set-game"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = ["pygame-ce"]

[project.optional-dependencies]
dev = ["pytest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
